Add process to cgroup before init process creation
